general stage reorg
