+++ examples/abp_pcap_detection/run.py
@@ -23,8 +23,8 @@ from morpheus.cli import str_to_file_type
 from morpheus.config import Config
 from morpheus.config import CppConfig
 from morpheus.config import PipelineModes
-from morpheus.stages.general_stages import AddClassificationsStage
-from morpheus.stages.general_stages import MonitorStage
+from morpheus.stages.general.general_stages import AddClassificationsStage
+from morpheus.stages.general.general_stages import MonitorStage
 from morpheus.stages.inference.inference_triton import TritonInferenceStage
 from morpheus.stages.input.file_source_stage import FileSourceStage
 from morpheus.stages.postprocess.serialize_stage import SerializeStage
