+++ morpheus/stages/general/trigger_stage.py
@@ -1,0 +1,82 @@
+# Copyright (c) 2021-2022, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import logging
+import typing
+from functools import reduce
+
+import cupy as cp
+import neo
+from neo.core import operators as ops
+from tqdm import TMonitor
+from tqdm import TqdmSynchronisationWarning
+from tqdm import tqdm
+
+import cudf
+
+import morpheus._lib.stages as neos
+from morpheus.config import Config
+from morpheus.config import CppConfig
+from morpheus.messages import MessageMeta
+from morpheus.messages import MultiMessage
+from morpheus.messages import MultiResponseProbsMessage
+from morpheus.pipeline import Stage
+from morpheus.pipeline.pipeline import SinglePortStage
+from morpheus.pipeline.pipeline import StreamPair
+from morpheus.utils.logging import deprecated_stage_warning
+
+logger = logging.getLogger(__name__)
+
+class TriggerStage(SinglePortStage):
+    """
+    This stage will buffer all inputs until the source stage is complete. At that point all messages
+    will be dumped into downstream stages. Useful for testing performance of one stage at a time.
+
+    Parameters
+    ----------
+    c : `morpheus.config.Config`
+        Pipeline configuration instance.
+
+    """
+
+    def __init__(self, c: Config):
+        super().__init__(c)
+
+    @property
+    def name(self) -> str:
+        return "trigger"
+
+    def accepted_types(self) -> typing.Tuple:
+        """
+        Accepted input types for this stage are returned.
+
+        Returns
+        -------
+        typing.Tuple
+            Accepted input types.
+
+        """
+        return (typing.Any, )
+
+    def _build_single(self, seg: neo.Segment, input_stream: StreamPair) -> StreamPair:
+
+        # Store all messages until on_complete is called and then push them
+        def node_fn(input: neo.Observable, output: neo.Subscriber):
+
+            input.pipe(ops.to_list(), ops.flatten()).subscribe(output)
+
+        node = seg.make_node_full(self.unique_name, node_fn)
+        seg.make_edge(input_stream[0], node)
+
+        return node, input_stream[1]
