+++ tests/benchmarks/test_bench_monitor_stage.py
@@ -23,7 +23,7 @@ import cudf
 from morpheus.config import Config
 from morpheus.config import CppConfig
 from morpheus.config import PipelineModes
-from morpheus.stages.general_stages import MonitorStage
+from morpheus.stages.general.general_stages import MonitorStage
 from morpheus.pipeline.pipeline import LinearPipeline
 from morpheus.stages.preprocess.preprocessing import DeserializeStage
 from morpheus.utils.logging import configure_logging
