+++ tests/test_add_classifications_stage.py
@@ -19,7 +19,7 @@ from unittest import mock
 import cupy as cp
 import pytest
 
-from morpheus.stages.general_stages import AddClassificationsStage
+from morpheus.stages.general.general_stages import AddClassificationsStage
 
 
 def test_constructor(config):
