+++ tests/test_phishing.py
@@ -22,8 +22,8 @@ import pytest
 
 from morpheus.config import PipelineModes
 from morpheus.pipeline import LinearPipeline
-from morpheus.stages.general_stages import AddClassificationsStage
-from morpheus.stages.general_stages import MonitorStage
+from morpheus.stages.general.general_stages import AddClassificationsStage
+from morpheus.stages.general.general_stages import MonitorStage
 from morpheus.stages.inference.inference_triton import TritonInferenceStage
 from morpheus.stages.input.file_source_stage import FileSourceStage
 from morpheus.stages.postprocess.serialize_stage import SerializeStage
